+++ survey/application/src/test/java/me/nalab/survey/application/service/create/SurveyCreateServiceTest.java
@@ -23,13 +23,14 @@ import me.nalab.survey.application.TestIdGenerator;
 import me.nalab.survey.application.common.dto.SurveyDto;
 import me.nalab.survey.application.exception.TargetDoesNotExistException;
 import me.nalab.survey.application.port.in.web.CreateSurveyUseCase;
-import me.nalab.survey.application.port.out.persistence.CreateSurveyPort;
-import me.nalab.survey.application.port.out.persistence.FindTargetPort;
+import me.nalab.survey.application.port.out.persistence.LatestSurveyIdFindPort;
+import me.nalab.survey.application.port.out.persistence.SurveyCreatePort;
+import me.nalab.survey.application.port.out.persistence.TargetExistCheckPort;
 import me.nalab.survey.domain.exception.IdAlreadyGeneratedException;
 
 @ExtendWith(SpringExtension.class)
-@ContextConfiguration(classes = {CreateSurveyService.class, TestIdGenerator.class})
-class CreateSurveyServiceTest {
+@ContextConfiguration(classes = {SurveyCreateService.class, TestIdGenerator.class})
+class SurveyCreateServiceTest {
 
 	@Autowired
 	private CreateSurveyUseCase createSurveyUseCase;
@@ -38,10 +39,13 @@ class CreateSurveyServiceTest {
 	private TestIdGenerator idGenerator;
 
 	@MockBean
-	private CreateSurveyPort createSurveyPort;
+	private SurveyCreatePort createSurveyPort;
 
 	@MockBean
-	private FindTargetPort findTargetPort;
+	private TargetExistCheckPort findTargetPort;
+
+	@MockBean
+	private LatestSurveyIdFindPort latestSurveyIdFindPort;
 
 	@ParameterizedTest
 	@MethodSource("surveyDtoLargeNullIdSources")
@@ -83,6 +87,10 @@ class CreateSurveyServiceTest {
 		// then
 		assertThrows(TargetDoesNotExistException.class, () -> createSurveyUseCase.createSurvey(targetId, null));
 	}
+
+	@Test
+	@DisplayName("최근 생성된 id 조회 성공 테스트")
+	voㅑㅇ
 
 	private static Stream<SurveyDto> surveyDtoLargeNullIdSources() {
 		RandomSurveyDtoFixture.initGenerator();
